requirements.txt
streamlit>=1.28.0
python-pptx>=0.6.21
openai-whisper>=20231117
